+++ src/components/organize/TagsManager/index.tsx
@@ -1,53 +1,13 @@
 import { useState } from 'react';
-import { Box, Switch, Tooltip, Typography } from '@material-ui/core';
+import { Box, Typography } from '@material-ui/core';
 import { FormattedMessage, useIntl } from 'react-intl';
 
 import ZetkinSection from 'components/ZetkinSection';
 import { ZetkinTag } from 'types/zetkin';
 
-const TagChip: React.FunctionComponent<{ tag: ZetkinTag }> = ({ tag }) => {
-  return (
-    <Tooltip arrow title={tag.description}>
-      <Box
-        bgcolor={tag.color || '#e1e1e1'}
-        borderRadius="18px"
-        fontSize={13}
-        px={2}
-        py={0.7}
-      >
-        {tag.title}
-      </Box>
-    </Tooltip>
-  );
-};
-
-const GroupToggle: React.FunctionComponent<{
-  checked?: boolean;
-  onChange: () => void;
-}> = ({ checked, onChange }) => {
-  const intl = useIntl();
-  return (
-    <Box alignItems="center" display="flex">
-      <Typography variant="body2">
-        {intl.formatMessage({ id: 'misc.tags.tagsManager.groupTags' })}
-      </Typography>
-      <Switch
-        checked={checked}
-        color="primary"
-        data-testid="TagsManager-groupToggle"
-        name="Tags"
-        onChange={onChange}
-      />
-    </Box>
-  );
-};
-
-interface TagsGroups {
-  [key: string]: {
-    tags: ZetkinTag[];
-    title: string;
-  };
-}
+import { groupTags } from './utils';
+import GroupToggle from './GroupToggle';
+import TagsList from './TagsList';
 
 const TagsManager: React.FunctionComponent<{
   appliedTags: ZetkinTag[];
@@ -55,28 +15,12 @@ const TagsManager: React.FunctionComponent<{
   const intl = useIntl();
 
   const [isGrouped, setIsGrouped] = useState(false);
-  const groups: TagsGroups = appliedTags.reduce((acc: TagsGroups, tag) => {
-    // Add to ungrouped tags list
-    if (!tag.group) {
-      return {
-        ...acc,
-        ungrouped: {
-          tags: acc['ungrouped'] ? [...acc['ungrouped'].tags, tag] : [tag],
-          title: intl.formatMessage({
-            id: 'misc.tags.tagsManager.ungroupedHeader',
-          }),
-        },
-      };
-    }
-    // Add to tags list for group
-    return {
-      ...acc,
-      [tag.group.id]: {
-        tags: acc[tag.group.id] ? [...acc[tag.group.id].tags, tag] : [tag],
-        title: tag.group.title,
-      },
-    };
-  }, {});
+  const groupedTags = groupTags(
+    appliedTags,
+    intl.formatMessage({
+      id: 'misc.tags.tagsManager.ungroupedHeader',
+    })
+  );
 
   return (
     <ZetkinSection
@@ -89,42 +33,14 @@ const TagsManager: React.FunctionComponent<{
       title={intl.formatMessage({ id: 'misc.tags.tagsManager.title' })}
     >
       <Box>
-        {/* Tags List */}
-        {appliedTags.length > 0 && (
-          <>
-            {isGrouped ? (
-              // Grouped list
-              <>
-                {Object.entries(groups).map(([id, group], i) => (
-                  <Box key={i} mb={1}>
-                    <Typography variant="overline">{group.title}</Typography>
-                    <Box
-                      data-testid={`TagsManager-groupedTags-${id}`}
-                      display="flex"
-                      flexWrap="wrap"
-                      style={{ gap: 8 }}
-                    >
-                      {group.tags.map((tag, i) => {
-                        // Tag Chip
-                        return <TagChip key={i} tag={tag} />;
-                      })}
-                    </Box>
-                  </Box>
-                ))}
-              </>
-            ) : (
-              // Flat list
-              <Box display="flex" flexWrap="wrap" style={{ gap: 8 }}>
-                {appliedTags.map((tag, i) => {
-                  // Tag Chip
-                  return <TagChip key={i} tag={tag} />;
-                })}
-              </Box>
-            )}
-          </>
-        )}
-        {/* If no tags */}
-        {appliedTags.length === 0 && (
+        {appliedTags.length > 0 ? (
+          <TagsList
+            groupedTags={groupedTags}
+            isGrouped={isGrouped}
+            tags={appliedTags}
+          />
+        ) : (
+          // If no tags
           <Typography>
             <FormattedMessage id="misc.tags.tagsManager.noTags" />
           </Typography>
