+++ src/components/organize/TagsManager/types.ts
@@ -1,0 +1,8 @@
+import { ZetkinTag } from 'types/zetkin';
+
+export interface TagsGroups {
+  [key: string]: {
+    tags: ZetkinTag[];
+    title: string;
+  };
+}
